tagilmo.vision and tagilmo.agenttools setup added + visualizer refactored
