+++ examples/5_lumberjack3D.py
@@ -1,17 +1,13 @@
-import math
 import torch
 import random
-import threading
-import cv2
 import numpy
-import os
 from time import sleep, time
 import tagilmo.utils.mission_builder as mb
 from examples.minelogy import Minelogy
-from examples.agent import TAgent
+from tagilmo.agenttools.agent import TAgent
 import logging
 from examples.log import setup_logger
-from examples.vis import Visualizer
+from tagilmo.vision.vis import Visualizer
 
 from tagilmo.utils.mathutils import *
