+++ examples/7_explorer.py
@@ -1,16 +1,9 @@
-import logging
 from examples.log import setup_logger
-from time import sleep, time
-import tagilmo.utils.mission_builder as mb
-from tagilmo.utils.mathutils import *
-import math
-from random import random
 
-from examples.vis import Visualizer
+from tagilmo.vision.vis import Visualizer
 from examples.minelogy import Minelogy
-from examples.goal import *
 from examples.skills import *
-from examples.agent import TAgent
+from tagilmo.agenttools.agent import TAgent
 
 class StaticKnowledge:
