+++ tagilmo/vision/common.py
