+++ tagilmo/vision/goodpoint.py
