+++ tagilmo/vision/pyramidpooling.py
